CompetitiveOffer refactored

+++ src/main/java/nure/abittour/dto/SubjectCoefDTO.java
@@ -3,9 +3,11 @@ package nure.abittour.dto;
 import lombok.Data;
 import nure.abittour.model.enums.Subject;
 
+import java.math.BigDecimal;
+
 @Data
 public class SubjectCoefDTO {
     private Long id;
     private String subject;
-    private Double coefficient;
+    private BigDecimal coefficient;
 }
